# application/services/custom_services/users/src/models/usersen.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const usersenSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   firstname: String,
   lastname: String,
   email: String,
   userpassword: String
})

const usersenModel = mongoose.model('usersen', usersenSchema, 'usersen');
export default usersenModel;